LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-hr-element-0/setting-overflow-visible-expected.txt


FAIL control assert_equals: hr.clientHeight expected 50 but got 0
FAIL overflow: visible assert_equals: hr.offsetLeft expected 0 but got 50
